refactor names added added more locators in the BasicForm
