+++ src/main/java/com/jalasoft/sdfc/pages/AssetsPage.java
@@ -6,7 +6,7 @@
  */
 package com.jalasoft.sdfc.pages;
 
-import com.jalasoft.sdfc.pages.common.BasicNewItemForm;
+import com.jalasoft.sdfc.pages.common.BasicForm;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
@@ -15,9 +15,9 @@ public class AssetsPage extends AbstractPage {
     @FindBy(xpath = "//ul[contains(@class, 'branding-actions')]/descendant::a/div[contains(text(), 'New')]")
     private WebElement newItemButton;
 
-    public BasicNewItemForm clickNewItemButton() {
+    public BasicForm clickNewItemButton() {
         action.click(newItemButton);
 
-        return new BasicNewItemForm();
+        return new BasicForm();
     }
 }
